refactor: make ParseOptions limit fields independent and Int-typed
expandedLength was a hidden nullable override that silently followed inputLength
whenever it was not set explicitly, which broke value substitutability: two
options with equal observable fields could carry different hidden state. Make
expandedLength an ordinary field defaulting to its own ResourceLimit bound, so
every ParseOptions field is independent and equality/hashing fold over exactly
the five visible fields.

Type the ResourceLimit defaults and LimitExceeded's observed/max as Int to match
InputTooLong and the length counts they compare against, removing the scattered
Int/Long conversions the mismatch forced.

+++ kuri/src/commonMain/kotlin/org/dexpace/kuri/error/UriParseError.kt
@@ -154,8 +154,8 @@ public sealed interface UriParseError {
      */
     public data class LimitExceeded(
         public val limit: ResourceLimit,
-        public val observed: Long,
-        public val max: Long,
+        public val observed: Int,
+        public val max: Int,
     ) : UriParseError
 
     /**
